third_party/WebKit/LayoutTests/editing/execCommand/create-list-with-hr.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
// This test pushes a horizontal rule into an unordered list with
// InsertUnorderedList. This demonstrates what might be a bug:
//  the horizontal rule appears before the list marker in the render tree.
test(() => assert_selection(
    '<div contenteditable>|<hr></div>',
    'insertUnorderedList',
    '<div contenteditable><ul><li>|<hr></li></ul></div>'),
    'Push HR into unordered list');
</script>
